backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
djangorestframework
djangorestframework_simplejwt
django-cors-headers
djoser==2.3.1
idna==3.10
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
# Custom users app for Djoser serializer